# Bugfix for waypoint turnrate