@JohnAFernandez 2021-06-14
Any release tested against 20.2 will be based on this behavior, so we can't put it into a point fix.