@Goober5000 2021-06-14
On the contrary, I would say that this fixes a major regression in behavior and therefore should be included.